move production code out of test, add ask for rec link, correct claim/file id
