+++ app/recommendations/[id]/RecommandationForm/Form.tsx
@@ -18,7 +18,7 @@ import { useStepContext } from '../../../credentialForm/form/StepContext'
 import { GoogleDriveStorage, saveToGoogleDrive } from '@cooperation/vc-storage'
 import { createDID, signCred } from '../../../utils/signCred'
 import { useSession } from 'next-auth/react'
-import ComprehensiveClaimDetails from '../../../test/[id]/ComprehensiveClaimDetails'
+import ComprehensiveClaimDetails from '../../../view/[id]/ComprehensiveClaimDetails'
 
 interface FormProps {
   fullName: string
