+++ app/test/[id]/page.tsx
@@ -2,7 +2,7 @@
 
 import React, { useState } from 'react'
 import { useParams } from 'next/navigation'
-import ComprehensiveClaimDetails from './ComprehensiveClaimDetails'
+import ComprehensiveClaimDetails from '../../view/[id]/ComprehensiveClaimDetails'
 const ClaimPage = () => {
   const params = useParams()
   const id =
